refactor(project): restructure package hierarchy for better organization
Reorganize package structure to align with domain-specific responsibilities. Key changes include:
- Moved `infra` contents to `adapters` and `util` for clearer separation of concerns.
- Renamed `ui` package to `presentation`.
- Adjusted imports and module descriptors to reflect new structure.

This improves maintainability, readability, and coherence of the codebase.
