+++ src/main/java/br/com/gzlabs/gzassist/adapters/GlobalHotkeyBinder.java
@@ -1,4 +1,4 @@
-package br.com.gzlabs.gzassist.infra;
+package br.com.gzlabs.gzassist.adapters;
 
 import br.com.gzlabs.gzassist.core.HotkeyBinder;
 import br.com.gzlabs.gzassist.errors.HotkeyException;
