+++ src/main/java/br/com/gzlabs/gzassist/presentation/UiEvent.java
@@ -1,4 +1,4 @@
-package br.com.gzlabs.gzassist.ui;
+package br.com.gzlabs.gzassist.presentation;
 
 public final class UiEvent {
