+++ src/main/resources/br/com/gzlabs/gzassist/presentation/home-view.fxml
@@ -6,7 +6,7 @@
 
 <?import javafx.scene.control.Button?>
 <VBox alignment="CENTER" spacing="20.0" xmlns:fx="http://javafx.com/fxml"
-      fx:controller="br.com.gzlabs.gzassist.ui.HomeController">
+      fx:controller="br.com.gzlabs.gzassist.presentation.HomeController">
     <padding>
         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
     </padding>
